requirements.txt
# TikTok Research System Dependencies

# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
numpy>=1.24.0

# Data processing
python-dateutil>=2.8.0
pytz>=2023.3

# Configuration management
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging and monitoring
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Scheduling
schedule>=1.2.0

# HTTP client enhancements
urllib3>=2.0.0
certifi>=2023.0.0

# JSON handling
jsonschema>=4.17.0

# Progress bars
tqdm>=4.65.0

# System monitoring
psutil>=5.9.0

